+++ spinta/types/config.py
@@ -66,6 +66,7 @@ def load(context: Context, config: Config) -> Config:
     config.env = rc.get('env')
     config.docs_path = rc.get('docs_path', default=None)
     config.always_show_id = rc.get('always_show_id', default=False)
+    config.root = rc.get('root', default=None)
 
     return config